+++ app/models/student.rb
@@ -27,15 +27,18 @@
 #  created_at             :datetime         not null
 #  updated_at             :datetime         not null
 #  group_id               :integer
+#  organization_id        :integer          not null
 #  profile_image_id       :integer
 #
 # Indexes
 #
 #  index_students_on_group_id          (group_id)
+#  index_students_on_organization_id   (organization_id)
 #  index_students_on_profile_image_id  (profile_image_id)
 #
 # Foreign Keys
 #
+#  fk_rails_...          (organization_id => organizations.id)
 #  fk_rails_...          (profile_image_id => student_images.id)
 #  students_group_id_fk  (group_id => groups.id)
 #
@@ -49,9 +52,11 @@ class Student < ApplicationRecord
 
   validates :first_name, :last_name, :dob, :gender, presence: true
   validates :mlid, uniqueness: { scope: :group_id }, length: { maximum: 3 }
+  validate :group_belongs_to_same_organization
 
   enum :gender, { M: 'male', F: 'female', NB: 'nonbinary' }
 
+  belongs_to :organization
   belongs_to :group
   has_many :grades, dependent: :restrict_with_error
   has_many :student_images, dependent: :restrict_with_error
@@ -71,10 +76,6 @@ class Student < ApplicationRecord
     "#{last_name}, #{first_name}"
   end
 
-  def organization
-    Organization.joins(chapters: :groups).find_by('groups.id = ?', group_id)
-  end
-
   def age
     now = Time.now.utc.to_date
     now.year - dob.year - (now.month > dob.month || (now.month == dob.month && now.day >= dob.day) ? 0 : 1)
@@ -91,5 +92,13 @@ class Student < ApplicationRecord
      :guardian_name, :guardian_occupation, :guardian_contact, :family_members, :health_insurance,
      :health_issues, :hiv_tested, :name_of_school, :school_level_completed, :year_of_dropout,
      :reason_for_leaving, :notes, :organization_id, :profile_image_id, { student_images_attributes: [:image], student_tags_attributes: [:tag_id, :student_id, :_destroy] }]
+  end
+
+  private
+
+  def group_belongs_to_same_organization
+    if group.chapter.organization_id != organization_id
+      errors.add(:group, I18n.t(:student_and_group_in_different_organizations))
+    end
   end
 end