+++ app/serializers/student_serializer.rb
@@ -27,15 +27,18 @@
 #  created_at             :datetime         not null
 #  updated_at             :datetime         not null
 #  group_id               :integer
+#  organization_id        :integer          not null
 #  profile_image_id       :integer
 #
 # Indexes
 #
 #  index_students_on_group_id          (group_id)
+#  index_students_on_organization_id   (organization_id)
 #  index_students_on_profile_image_id  (profile_image_id)
 #
 # Foreign Keys
 #
+#  fk_rails_...          (organization_id => organizations.id)
 #  fk_rails_...          (profile_image_id => student_images.id)
 #  students_group_id_fk  (group_id => groups.id)
 #